absorb(upstream): Bun + servo dual-window wave — 51 upstream fixes + 8 native fixes
Bun window e6978042a..01c4e2fd6 (309 commits: 12 absorbed, 15 already-contained,
253 batched N/A, 29 verified-adjudicated):
- usockets C: b9f509020 (eof/error half-open hang), a35696478 (closed-socket
  poll-error skip), 4199361ed (TLS RST terminate, node resetAndDestroy parity)
- uws: cdd2d05c6 (close-delimited Content-Length into body), 6b57f807f
  (destroy() writes Connection:close after headers_open)
- rust: 72e37255c (fetch TCP keepalive once per conn), 8226b3d6e (clamp loses
  lower bound), 8f8695f4c (rgba() missing delim), 20d04673b (brace slot u16
  overflow), a53ac0c0e (watchlist eviction path leak), 668caf3da (>=2GiB
  source rejected before parse, P0), 4673c48bb (dead dynamic import chunk)
- 8bc4d2a88 (fallible allocation hardening, 18 files: UninitBuf/sb_try_reserve,
  OOM error surfaces zstd/zlib/brotli, zero-fill elimination base64/paths/sys/
  http/uws/bundler; data_url historical transcription divergence corrected)
- csrc mirror protocol restored: 0825a8b3f (ssl deferred close) enters the
  compiled tree for the first time + :981 void-return transcription fixed
  (clang 23 -Wreturn-mismatch)

servo window 4842b770e..0d579bd5a (109 commits: 38 absorbed, 16 already-contained,
46 batched N/A, 9 adjudicated):
- P1: 52bd46cc0, 27775d1d9, 113679f9f, b86633c99, 67d568715, 8462c2bb0,
  d156d4dde, 79ac6c56f
- P2: 4a2e02343, 0753df091, 6da8677fd, 0d558ccb4, 4f616abe0, 808e65262,
  6ee336ce8 (MouseButton Left->Primary + bao_browser page.rs adaptation),
  544c99bc9 (script/drag), 80b2c05fb, 7f8f0862d, ea54c00a1 (about:blank
  replace, CDP evaluate semantics assessed), 4a599939e, 88497f86e (stealth
  impact assessed: double-gated, dormant by default)
- P3: 7fcdc22f6, a57e1ff10, 27ccefba3, 09a6a4fe4, 8f652c303, 49065a681,
  66bac8a96, 6de1edc66, 68813fb12, 9e00f85c8 + a986ae109 (text-transform
  math-auto/full-width via bao-stylo fork PR 442), a3d9ed9f0, 9701eef04,
  ed93ab701, ddfc66dec, e7b419402
- BCE patch anchors fully preserved (script_runtime x4, constellation x6,
  script_thread x4, shared script lib x1, resource_thread x1)

Bao-native fixes (verify-bun-b confirmed shapes + dns deep-dive):
- dns/vm Node parity: 3rd-arg callback (c4618cd36-shape), vm options
  validateObject (60f6e186c-shape), Script ctor syntax check (1b881a9d2-shape)
- dns native P0: ares_free_hostent double-free SIGSEGV (NS/CNAME/PTR live
  records) + 4 semantics fixes (NS via ares_parse_ns_reply, PTR via
  get_host_by_addr, SOA use-after-free snapshot-before-free, ENOTFOUND) —
  Node 24.5.0 oracle A/B verified byte-for-byte

Verification: cargo check -p bao-browser -p bun_runtime -p bun_uws_sys EXIT=0;
bao_uloop 45/45 (hangup/paused_eof/write_rearm semantics); bun_http 25/25
isolated (tail flake is environmental, excluded by no-0825 control experiment);
dns 3/3; bun_collections 35/35; vm/dns scoped 12/12.

Residual ledger: fetch unbounded body buffering (7da5a09fa stronger form,
needs architecture decision), 3d77be2fc (blocked on mozjs fork throw_*_safe
API), stylo fork rev sync, CookieStoreManager scope, bun_http tail-flake infra.

+++ src/shell_parser/braces.rs
@@ -654,7 +654,7 @@ pub fn expand(
     contains_nested: bool,
 ) -> Result<(), ExpandError> {
     check_brace_group_count(tokens)?;
-    let mut out_key_counter: u16 = 1;
+    let mut out_key_counter: usize = 1;
     if !contains_nested {
         let expansions_table = build_expansion_table_alloc(tokens)?;
 
@@ -688,8 +688,8 @@ pub fn expand(
 unsafe fn expand_nested(
     root: *mut ast::Group,
     out: &mut [Vec<u8>],
-    out_key: u16,
-    out_key_counter: &mut u16,
+    out_key: usize,
+    out_key_counter: &mut usize,
     start: u32,
 ) -> Result<(), ExpandError> {
     // SAFETY: see fn doc comment — raw-pointer derefs mirror Zig pointer semantics;
@@ -713,7 +713,7 @@ unsafe fn expand_nested(
 
             match &(*root).atoms {
                 ast::GroupAtoms::Single(ast::Atom::Text(txt)) => {
-                    out[usize::from(out_key)].extend_from_slice(txt.slice());
+                    out[out_key].extend_from_slice(txt.slice());
                     if !(*root).bubble_up.is_null() {
                         let bubble_up = (*root).bubble_up;
                         let next = (*root).bubble_up_next.unwrap();
@@ -728,11 +728,11 @@ unsafe fn expand_nested(
                     return Ok(());
                 }
                 ast::GroupAtoms::Single(ast::Atom::Expansion(expansion)) => {
-                    let length = out[usize::from(out_key)].len();
+                    let length = out[out_key].len();
                     // PORT NOTE: reshaped for borrowck — snapshot prefix once; Zig re-sliced
                     // out[out_key].items[0..length] each iteration (same bytes).
                     // PERF(port): extra Vec alloc for prefix snapshot — profile if hot.
-                    let prefix: Vec<u8> = out[usize::from(out_key)][..length].to_vec();
+                    let prefix: Vec<u8> = out[out_key][..length].to_vec();
                     let variants = expansion.variants;
                     let variants_len = variants.len();
                     for j in 0..variants_len {
@@ -743,7 +743,7 @@ unsafe fn expand_nested(
                             out_key
                         } else {
                             let new_key = *out_key_counter;
-                            out[usize::from(new_key)].extend_from_slice(&prefix);
+                            out[new_key].extend_from_slice(&prefix);
                             *out_key_counter += 1;
                             new_key
                         };
@@ -776,13 +776,13 @@ unsafe fn expand_nested(
             let atom: &ast::Atom = &(*many)[i_];
             match atom {
                 ast::Atom::Text(txt) => {
-                    out[usize::from(out_key)].extend_from_slice(txt.slice());
+                    out[out_key].extend_from_slice(txt.slice());
                 }
                 ast::Atom::Expansion(expansion) => {
-                    let length = out[usize::from(out_key)].len();
+                    let length = out[out_key].len();
                     // PORT NOTE: reshaped for borrowck — see above.
                     // PERF(port): extra Vec alloc for prefix snapshot — profile if hot.
-                    let prefix: Vec<u8> = out[usize::from(out_key)][..length].to_vec();
+                    let prefix: Vec<u8> = out[out_key][..length].to_vec();
                     let variants = expansion.variants;
                     let variants_len = variants.len();
                     for j in 0..variants_len {
@@ -793,7 +793,7 @@ unsafe fn expand_nested(
                             out_key
                         } else {
                             let new_key = *out_key_counter;
-                            out[usize::from(new_key)].extend_from_slice(&prefix);
+                            out[new_key].extend_from_slice(&prefix);
                             *out_key_counter += 1;
                             new_key
                         };
@@ -821,8 +821,8 @@ fn expand_flat(
     tokens: &[Token],
     expansion_table: &[ExpansionVariant],
     out: &mut [Vec<u8>],
-    out_key: u16,
-    out_key_counter: &mut u16,
+    out_key: usize,
+    out_key_counter: &mut usize,
     depth_: u8,
     start: usize,
     end: usize,
@@ -836,7 +836,7 @@ fn expand_flat(
     for atom in tokens[start..end].iter() {
         match atom {
             Token::Text(txt) => {
-                out[usize::from(out_key)].extend_from_slice(txt.slice());
+                out[out_key].extend_from_slice(txt.slice());
             }
             Token::Close => {
                 depth -= 1;
@@ -851,16 +851,16 @@ fn expand_flat(
                     [usize::from(expansion_variants.idx)..usize::from(expansion_variants.end)];
                 let skip_over_idx = variants[variants.len() - 1].end();
 
-                let starting_len = out[usize::from(out_key)].len();
+                let starting_len = out[out_key].len();
                 // PORT NOTE: reshaped for borrowck — snapshot prefix once.
                 // PERF(port): extra Vec alloc for prefix snapshot — profile if hot.
-                let prefix: Vec<u8> = out[usize::from(out_key)][..starting_len].to_vec();
+                let prefix: Vec<u8> = out[out_key][..starting_len].to_vec();
                 for (i, variant) in variants.iter().enumerate() {
                     let new_key = if i == 0 {
                         out_key
                     } else {
                         let new_key = *out_key_counter;
-                        out[usize::from(new_key)].extend_from_slice(&prefix);
+                        out[new_key].extend_from_slice(&prefix);
                         *out_key_counter += 1;
                         new_key
                     };
